tests: Move `GET /crates/:name/dependencies` test to `krate::dependencies` module
